requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
Werkzeug==3.1.3
azure-storage-blob==12.19.1
python-dotenv==1.0.1

